[+] remove list

+++ public/js/ngApp/components/dashboard-resources-contactLists/ContactListsController.js
@@ -2,8 +2,8 @@
   'use strict';
   angular.module('KliikoApp').controller('ContactListController', ContactListController);
 
-  ContactListController.$inject = ['domServices','contactListServices', 'dbg', 'messenger', '$timeout'];
-  function ContactListController(domServices, contactListServices, dbg, messenger, $timeout) {
+  ContactListController.$inject = ['domServices','contactListServices', 'dbg', 'messenger', '$timeout', '$scope'];
+  function ContactListController(domServices, contactListServices, dbg, messenger, $timeout, $scope) {
     dbg.log2('#ContactListController  started');
     var vm =  this;
 
@@ -13,13 +13,18 @@
     vm.newContact = {};
     vm.newList = {};
     vm.modalErrors = {};
+    vm.allSelected = false;
 
     vm.listItemClickHandle = listItemClickHandle;
     vm.addContactManual = addContactManual;
     vm.createContact = createContact;
+    vm.selectAll = selectAll;
     vm.removeContacts = removeContacts;
+    vm.massDelete = massDelete;
     vm.addNewList = addNewList;
     vm.submitNewList = submitNewList;
+    vm.removeList = removeList;
+
 
     init();
 
@@ -32,21 +37,29 @@
           vm.activeListIndex = 0;
           vm.selectedListMembers = vm.lists[vm.activeListIndex].members;
 
-          prepareListControls();
+          prepareSelectedListMembersControls();
         }
       });
+    }
 
-      function prepareListControls() {
-        for (var i = 0, len = vm.lists.length; i < len ; i++) {
-          vm.lists[i].listControls = {
-            checked: false
-          }
-        }
+
+    function prepareSelectedListMembersControls() {
+      if (!vm.selectedListMembers) return;
+
+      for (var i = 0, len = vm.selectedListMembers.length; i < len ; i++) {
+        if (!vm.selectedListMembers[i].selected) vm.selectedListMembers[i].selected = false;
       }
     }
 
-    function listItemClickHandle(item, index){
+    /**
+     * Go to selected list
+     * @param index {number} - index in vm.lists array
+     */
+    function listItemClickHandle(index){
       vm.activeListIndex = index;
+      vm.selectedListMembers = vm.lists[vm.activeListIndex].members;
+
+      prepareSelectedListMembersControls();
     }
 
     function addContactManual() {
@@ -70,10 +83,15 @@
           vm.newContact = {};
           newContact.id = res.id;
           vm.lists[vm.activeListIndex].membersCount++;
+          if (!vm.selectedListMembers) vm.selectedListMembers = [];
           vm.selectedListMembers.push(newContact);
+          prepareSelectedListMembersControls();
 
         },
         function(err) {
+          if (err.error) {
+            messenger.error(err.error.message);
+          }
           if (err.errors) {
             var e = err.errors;
             for (var i = 0, len = e.length; i < len ; i++) {
@@ -83,7 +101,6 @@
 
         }
       );
-
 
 
 
@@ -100,8 +117,24 @@
 
     }
 
+    function selectAll() {
+      vm.allSelected = !vm.allSelected;
+      for (var i = 0, len = vm.selectedListMembers.length; i < len ; i++) {
+        vm.selectedListMembers[i]._selected = vm.allSelected;
+      }
+    }
+
+    /**
+     * Remove contacts from list by given ids
+     * @param ids {number | [{numbers}]}
+     */
     function removeContacts(ids) {
-      console.log(ids);
+      if (!ids) return;
+      if (!angular.isArray(ids)) ids = [ids];
+
+      var confirmed = confirm('Are you sure?');
+      if (!confirmed) return;
+
       contactListServices.deleteUser(ids).then(
         function(res) {
 
@@ -118,26 +151,28 @@
           messenger.ok(message);
 
           // remove this user(s) from view
-          var index = [];
           for (var i = 0, len = ids.length; i < len ; i++) {
 
-            for (var j = 0, len = vm.selectedListMembers.length; j < len ; j++) {
-              if (vm.selectedListMembers[j].id == ids[i] ) index.push(j);
+            for (var j = 0; j < vm.selectedListMembers.length; j++) {
+              if (vm.selectedListMembers[j].id == ids[i]) {
+                vm.selectedListMembers.splice( j, 1 );
+                vm.lists[vm.activeListIndex].membersCount--
+              }
             }
 
           }
-          for (var i = 0, len = index.length; i < len ; i++) {
-            vm.selectedListMembers.splice( index[i], 1 );
-          }
-
-          vm.lists[vm.activeListIndex].membersCount--;
-
         },
-        function(err) {
-          messenger.error(err);
-
-        }
+        function(err) {  messenger.error(err); }
       );
+    }
+
+    function massDelete() {
+      var ids = [];
+      for (var i = 0, len = vm.selectedListMembers.length; i < len ; i++) {
+        if (vm.selectedListMembers[i]._selected === true) ids.push(vm.selectedListMembers[i].id);
+      }
+
+      removeContacts(ids);
     }
 
     function addNewList() {
@@ -166,6 +201,43 @@
 
       );
     }
+
+    function removeList() {
+      var confirmed = confirm('Are you sure?');
+      if (!confirmed) return;
+
+      var list = vm.lists[vm.activeListIndex];
+      contactListServices.deleteList( list.id ).then(
+        function(res) {
+          if (!res.success) {
+            dbg.error('#ContactListController > removeList > something wrong with res output');
+            messenger.error('There is an error while removing the list');
+            return;
+          }
+
+          //delete from view
+          vm.lists.splice(vm.activeListIndex, 1);
+
+          //select previous in array list
+          //debugger
+          vm.activeListIndex--;
+          if (vm.activeListIndex < 0) vm.activeListIndex = 0;
+          listItemClickHandle(vm.activeListIndex);
+
+          dbg.log('#ContactListController > removeList > success: List "'+ list.name + '" removed');
+          messenger.ok('List "'+ list.name +'" successfully removed');
+        },
+        function(err) {
+          messenger.error('There is an error while removing the list');
+          dbg.error('#ContactListController > removeList > error: ', err);
+        }
+      )
+    }
+
+
+
+
+
 
   }
 })();
